requirements.txt
python-dotenv
geopandas
fiona
numpy
scikit-learn
ipython
ipyleaflet
ipykernel
ipywidgets
jupyterlab
joblib
matplotlib
shapely
tqdm
gsutil
pyarrow
xgboost
earthengine-api
geemap
gcsfs
lonboard
duckdb>1.0.0
rasterio
google-cloud-storage
s3fs
tenacity
PyYAML
voila
